Setup defaults

+++ src/api/v1/rank_routes.py
@@ -29,8 +29,9 @@ from sqlalchemy.exc import IntegrityError, SQLAlchemyError # to catch db errors
 from flask_smorest import Blueprint, abort # type: ignore
 from uuid import UUID
 
-from src.api.models import RankModel # type: ignore
+from src.api.models import MemberModel, RankModel # type: ignore
 from src.api.schemas import MessageSchema, RankQueryArgsSchema, RankSchema
+from src.constants import DEFAULT_RANK
 
 from src.extensions import db
 
@@ -164,8 +165,16 @@ class RankByIdResource(MethodView):
         except ValueError:
             abort(400, message="Invalid rank id")
 
+        # Check rank exists if not return a 404
         rank = RankModel.query.get_or_404(data)
 
+        # If default rank, do not allow delete
+        if rank_id == str(DEFAULT_RANK):
+            abort(400, message="You cannot delete the default rank")
+
+        # Update any members who have that rank to the default rank
+        MemberModel.query.filter_by(rank_id=rank_id).update({'rank_id': DEFAULT_RANK}) # TODO: turn into protected rank and make a const
+    
         try:
             db.session.delete(rank)
             db.session.commit()
